bots/requirements.txt
aiogram==3.3.0
httpx
python-dotenv
